refactor: drop redundant int/int64 casts after strategy.ID is int64
Now that #7 made strategy.Strategy.ID and related fields int64, the
int(p.StrategyID) bridge casts and int64(s.ID) widening casts at call
sites are no longer needed — unconvert flagged 64 of them. Also
migrate the one remaining int-typed method on the strategies branch
(UpdateRules) to int64 for consistency, and switch a few strconv.Itoa
calls to strconv.FormatInt now that s.ID is int64.

+++ internal/email/proposals_test.go
@@ -62,7 +62,7 @@ func seedFixture(t *testing.T) (
 	if err != nil {
 		t.Fatalf("seed strategy: %v", err)
 	}
-	if err := sRepo.Verify(ctx, int64(s.ID)); err != nil {
+	if err := sRepo.Verify(ctx, s.ID); err != nil {
 		t.Fatalf("verify: %v", err)
 	}
 	got, _ := sRepo.GetByID(ctx, s.ID)
@@ -71,7 +71,7 @@ func seedFixture(t *testing.T) (
 		UserID:            uid,
 		Name:              "Quality Compounders",
 		StartingCapital:   decimal.NewFromInt(10000),
-		StrategyID:        int64(s.ID),
+		StrategyID:        s.ID,
 		StrategyVersionID: *got.CurrentVersionID,
 		Cadence:           strategy.CadenceQuarterly,
 	})
