+++ internal/strategy/repository_status_test.go
@@ -27,7 +27,7 @@ func TestVerify_FromDraftSucceeds(t *testing.T) {
 	repo := strategy.NewRepository(pool)
 	s := seedStrategy(t, pool, "VerifyDraft")
 
-	if err := repo.Verify(ctx, int64(s.ID)); err != nil {
+	if err := repo.Verify(ctx, s.ID); err != nil {
 		t.Fatalf("verify: %v", err)
 	}
 	got, _ := repo.GetByID(ctx, s.ID)
@@ -41,9 +41,9 @@ func TestVerify_FromVerifiedIsIdempotent(t *testing.T) {
 	ctx := context.Background()
 	repo := strategy.NewRepository(pool)
 	s := seedStrategy(t, pool, "VerifyTwice")
-	_ = repo.Verify(ctx, int64(s.ID))
+	_ = repo.Verify(ctx, s.ID)
 
-	if err := repo.Verify(ctx, int64(s.ID)); err != nil {
+	if err := repo.Verify(ctx, s.ID); err != nil {
 		t.Errorf("second verify should be a no-op, got err: %v", err)
 	}
 }
@@ -53,9 +53,9 @@ func TestVerify_FromArchivedFails(t *testing.T) {
 	ctx := context.Background()
 	repo := strategy.NewRepository(pool)
 	s := seedStrategy(t, pool, "VerifyArchived")
-	_ = repo.Archive(ctx, int64(s.ID))
+	_ = repo.Archive(ctx, s.ID)
 
-	err := repo.Verify(ctx, int64(s.ID))
+	err := repo.Verify(ctx, s.ID)
 	if !errors.Is(err, strategy.ErrInvalidStatusTransition) {
 		t.Errorf("err = %v, want ErrInvalidStatusTransition", err)
 	}
@@ -79,7 +79,7 @@ func TestArchive_FromAnyStateSucceeds(t *testing.T) {
 
 	// from draft
 	s1 := seedStrategy(t, pool, "ArchDraft")
-	if err := repo.Archive(ctx, int64(s1.ID)); err != nil {
+	if err := repo.Archive(ctx, s1.ID); err != nil {
 		t.Errorf("archive from draft: %v", err)
 	}
 	got1, _ := repo.GetByID(ctx, s1.ID)
@@ -89,13 +89,13 @@ func TestArchive_FromAnyStateSucceeds(t *testing.T) {
 
 	// from verified
 	s2 := seedStrategy(t, pool, "ArchVerified")
-	_ = repo.Verify(ctx, int64(s2.ID))
-	if err := repo.Archive(ctx, int64(s2.ID)); err != nil {
+	_ = repo.Verify(ctx, s2.ID)
+	if err := repo.Archive(ctx, s2.ID); err != nil {
 		t.Errorf("archive from verified: %v", err)
 	}
 
 	// already archived (idempotent)
-	if err := repo.Archive(ctx, int64(s1.ID)); err != nil {
+	if err := repo.Archive(ctx, s1.ID); err != nil {
 		t.Errorf("archive twice should be a no-op, got err: %v", err)
 	}
 }
